Hex encode binary data to and from the hardware. More inside.
So on NativeScript the http library has some pretty significant shortcomings,
specifically the refusal to send binary data in the POST. So this works around
that by temporarily hex encoding the data. We indicate our desire to talk hex
encoded by sending over the text/plain content type. Once we have a good idea of
the full scope of the communications related issues we'll be fixing this, and
only have to change this deviceQuery method.