Manually set ownerRef as an instance property on listener callback.
My rough theory is that behind the scenes NS is caching the class definition and
failing to use the new one that's extended here. One thing I noticed was if I
created an attribute 'id: id' on this listener class and logged the value of
listener.id after I new'd the listener the value was *always* the first value I
created.

So, basically the contents of the { } blocked that's passed to extend the first
time ends up being the one and only capture instance that's ever instantiated.
Another clue that this was happening was I tried adding a ctor parameter to pass
owner/id into and would get errors like this:

Failed resolving constructor for class
\'co.fitcom.fancydownloader.DownloadListenerUI_update_140_73_\' with 1
parameters. Check the number and type of arguments.

If you noticed the _update_140_73 suffix, that's the clue. This would *never*
change after the first call. So it seemed like that represented a behind the
scenes object that was being constructed by NS and never updated in future
calls, I'm willing to bet that inside whatever is happening in "extend" there's
something that checks for a key in some kind of cache and just returns whatever
is in the cache (in our case that would be the first object we passed in)

So my "fix" is to manually set ownerRef *after* I created an instance,
overwriting the value that was set form the first cached instance.

Or I'm totally wrong.