Some larger GPS syncing fixes.
A few things happening... sometimes the phone's location just takes
too long to come back and by then the tester/user has already visited
the StationDetails screen, so when the location does come in we aren't
necessarily going to get a chance to update the portal.

I'm generally not a fan of coupling things that appear to be
background updates to the user to a particular screen. So we should be
moving to performing these updates in the service layer, which this
commit starts to do.

When the phone's location does come in, that's when we update
connected stations and then if there's portal connection we update the
portal, behind the scenes.

I've also started moving some logic into classes.