Start to pull state handling logic into domain objects.
My goal is to make some areas more testable by decoupling code from
the complicated service layer we have now that largely depends on the
TNS modules and wrappers.

So this is the start of a domain layer, which is classes that touch a
very narrow portion of our services layer and produce very few promises.