Logging around tests... and a theory.
Locally, this passes but does so in a way that I'm skeptical of. For
example. The insertion isn't happening locally, for me but the test
still passes and from what I can tell there's nothing faulty with the
test that would cause a false positive like this. So, I'm gonna see
about fixing the test and we'll figure out why this was passing so
often.