Fully functional test for discovering a station.
This also tests to ensure we're adding data to the database that is
expected. Part of this process was cleaning up error messages and also
ensuring that catch's propagate the error. We need to be way more
careful about when we just swallow them, which I'm hoping writing
tests prevents us from getting lazy on.

Hopefully going to be fewer large changes now.