Simplifying the handling of subscriptions in StationManager
Having the 3 different kinds had kind of lost all of the original
benefit because we ended up publishing them at the same time, to
simplify thinking about these things we'll just publish one event. We
typically don't have that many subscribers, anyway and so the
performance benefits of publishing smaller, finer grained events
disappeared with that and the fact that we were publishing anyway.