Buffer loss events just in case they come right back.
I was noticing locally that my station would get a service lost event
and then quickly get a station found. What's weird is there's no
network traffic associated with these and the events seem to just be
spurious rumblings from Android.

What this does is buffer the loss event for 2s, and gives the found
event a chance to cancel that.