Fix some startup issues, defer app startup work until we have UI.
It seems, unknown to me that the $start call on Vue is blocking and so
many things weren't being setup at startup. This makes the move in the
right direction to do full app setup stuff once we have UI up so we
can show helpful messages, etc... It's not super pretty yet, though.