Pull clock into the state, normalize how we declare module states.
The problem with the getter was passing in the time. Because the state
wasn't actually being updated the Vue would be stale until a mutation
happened. So our choice is to either have the Vue layer requery the
state with the getter or just push the clock in as state and update
once a second. For now this is the predictable way so it's the winner.