Fix saving notes race condition.
This is due to the fact that we're not properly obeying Vuex semantics
here yet, so some changes were being lost, basic the object being
mutated was the state object that would get orphaned sometimes, and so
wasn't involved in the save. This sets the change on the actual local
objects we have before dispatching.

Still not great until we move to proper mutations here.