Fix iOS crash due to DropDown race. FK-2837
This was very strange and didn't happen all the time, only in certain
situations and so I don't really know what the problem is. Especially
since the data here is static. I thought it was data related, but
maybe the hangup is a speed related thing depending on the data.

A true fix might involve upgrading to the paid DropDown.

Anyway, delaying the initialize of the DropDown like this fixed the issue.