Fix jest tests after type script changes.
After the tsconfig changes things on the jest end went to shit, so
some changes had to happen. The 'wrappers' system was basically a poor
rendition of what we want from the __mocks__ concept, which came about
in my dead zone of java script.

So this starts to eliminate the indirection in wrappers and move tests
to a more cohesive language/syntax system.