Finally fix sudden iOS startup crash.
This was a pain. v8 was just crashing at startup, with no information at all.
So I've been basically A/B testing the code to try and narrow down the
offending part, starting with the application entrypoint.

Turns out that it's related to the Worker stuff, and big surprise there. I'm
gonna be able to narrow this down more, but wanted to confirm that this also
fixes things on the build server.