Respond to changes in connected WiFi network.
The idea here is that if we see the network change we can make some
assumptions. First is we probably need to check in on the stations
we've been talking to recently, the other is that if there's no
current wifi then we're no longer talking to who we were.

So right now we LOSE all stations when there's no longer a wifi
network and then we also do a nearby scan when we find ourselves on a
new one.

This can be loud in the logs, right now.