Retry twice until firmware startup order is fully released.
Sometimes we can try too soon, and the station isn't totally ready
yet. So this adds a very short delay and tries more than once. I also
made a change to the firmware to announce after the servers are
listening, so this is a stop gap until that's fully deployed.