Push active user to store, map to user from station user id.
This wires up the backend support for allowing us to pick the user to
use when talking to the portal about a particular station. Also cleans
up a ton of things here and there in that process.

Before we were doing a redundant update to the portal for station
data, this relies purely on the add station call.