fix: Don't allow station listing screens to pile up on backstack.
This fixes a nav related issue that was causing station detail screens
to stick around and was a bug in how we were handling navigating in
other tabs than the one we're on.

This brings back the event bus to change the selected tab when the
navigation code determines that's necessary, which is most of the time
since we're always in the tabbed layout for the most part.