fix: Forget station breaks with previous screens on the backstack.
The problem, as I can see things was that NS-Vue keeps previous pages
alive on the backstack for navigation and so when the station
disappears that causes errors that vue can't seem recover from. I
tried a few alternatives to this and none of them made me
happy. Ideally I think the other pages would play friendlier but as a
quick solution the navigation back to the listing screen and the async
delay so that the forget happens after seems like a quick and dirty
solution, esp as this isn't a common use case.